Sorted out the Notebook, canvas and mouse_wheel

+++ config.py
@@ -7,7 +7,7 @@ class Config(App):
 
     def mouse_scroll(self, event):
         if event.delta:
-            self.tasks_canvas.yview_scroll(-1 * (event.delta / 120), "units")
+            self.tasks_canvas.yview_scroll(int(-1 * (event.delta / 120)), "units")
         else:
             if event.num == 5:
                 move = 1
